Fix size_t warnings

+++ src/adaptor.cc
@@ -138,7 +138,7 @@ void delete_quic_encrypted_packet(QuicEncryptedPacket *packet) {
 
 IPAddressNumber *create_ip_address_number(unsigned char *ip_buf, size_t length) {
   IPAddressNumber *ip = new IPAddressNumber;
-  for (int i = 0; i < length; i++) {
+  for (size_t i = 0; i < length; i++) {
     ip->push_back(ip_buf[i]);
   }
   return ip;
@@ -155,7 +155,7 @@ IPEndPoint *create_ip_end_point(IPAddressNumber *ip, uint16_t port) {
 size_t ip_endpoint_ip_address(IPEndPoint *ip_end_point, void *address_buf) {
   const IPAddressNumber &ip = ip_end_point->address();
   const size_t size = ip.size();
-  for (int i = 0; i < size; i++) {
+  for (size_t i = 0; i < size; i++) {
     // XXX(serialx): Performance and safety issues here. Ensure address_buf is large enough for ipv6
     ((char *)address_buf)[i] = ip[i];
   }
